rest_server/requirements.txt
numpy == 1.25.1
scikit-learn == 1.3.0
django == 4.2.3
django_rest_framework == 0.1.0
tqdm